refactor(presets): extract shared PresetsModel from StationListModel
